Solución de los Application Context startup failures
Varios errores que impedían arrancar la aplicación

+++ src/main/java/io/github/dainadb/improplan/domain/event/dto/EventRequestDto.java
@@ -1,5 +1,6 @@
 package io.github.dainadb.improplan.domain.event.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 
@@ -24,12 +25,12 @@ public class EventRequestDto {
     private String description;
     private String placeName;
     private String address;
-    private Double latitude; //Estos datos tienen un valor por defecto de 0.0 porque el usuario (role-user) no introduce latitud ni longitud cuando crea la solicitud del evento. Estos datos los introducirá el administrador manualmente cuando publique el evento.
-    private Double longitude; //Estos datos tienen un valor por defecto de 0.0 porque el usuario (role-user) no introduce latitud ni longitud cuando crea la solicitud del evento. Estos datos los introducirá el administrador manualmente cuando publique el evento.
+    private BigDecimal latitude; //Estos datos tienen un valor por defecto de 0.0 porque el usuario (role-user) no introduce latitud ni longitud cuando crea la solicitud del evento. Estos datos los introducirá el administrador manualmente cuando publique el evento.
+    private BigDecimal longitude; //Estos datos tienen un valor por defecto de 0.0 porque el usuario (role-user) no introduce latitud ni longitud cuando crea la solicitud del evento. Estos datos los introducirá el administrador manualmente cuando publique el evento.
     private String image;
     private String infoUrl;
     private Boolean isFree;
-    private Double price;
+    private BigDecimal price;
     //private Boolean inTime; //Es un campo que por defecto será true ya que cuando se cree un evento las fechas serán futuras y lo modificará un método programado (scheduler).
     private Event.StatusType status; // Es un campo que por defecto será PENDING cuando se cree un evento. El administrador podrá cambiarlo a PUBLISHED o DISCARDED.
     private String municipalityName;
